Aligned ExtractId to extract a long, since MAL id can be just that.

+++ API/Entities/Scrobble/ScrobbleEvent.cs
@@ -15,7 +15,7 @@ public class ScrobbleEvent : IEntityDate
     public required ScrobbleEventType ScrobbleEventType { get; set; }
 
     public int? AniListId { get; set; }
-    public int? MalId { get; set; }
+    public long? MalId { get; set; }
 
 
     /// <summary>
